+++ geopandas/_vectorized.py
@@ -721,6 +721,23 @@ def interiors(data):
     return data
 
 
+def remove_repeated_points(data, tolerance=0.0):
+    if compat.USE_SHAPELY_20:
+        return shapely.remove_repeated_points(data, tolerance=tolerance)
+    if compat.USE_PYGEOS and compat.SHAPELY_GE_20:
+        warnings.warn(
+            "PyGEOS does not support concave_hull, and Shapely >= 2 is installed, "
+            "thus using Shapely and not PyGEOS to remove repeated points.",
+            stacklevel=4,
+        )
+        return shapely.remove_repeated_points(data, tolerance=tolerance)
+    else:
+        raise NotImplementedError(
+            f"shapely >= 2.0 is required, "
+            f"version {shapely.__version__} is installed"
+        )
+
+
 def representative_point(data):
     if compat.USE_PYGEOS:
         return pygeos.point_on_surface(data)